ChessDecoderNode + Fix Docker Display
Edit run_dev.sh
- Forward display from Docker (DISPLAY=0:0)
Edit workspace_entrypoint.py
- Install libv4l-dev beforebuild & install python-v4l2capture
- build&install OpenCV only for arrch64
- build & install "DOPE"
- res & build & install "module89"

+++ docker/scripts/workspace-entrypoint.sh
@@ -2,7 +2,10 @@
 
 # Build ROS dependency
 echo "source /opt/ros/foxy/setup.bash" >> ~/.bashrc
-source /opt/ros/foxy/setup.bash
+echo "alias res=\"cd /workspaces/isaac_ros-dev && rosdep install -i --from-path /workspaces/isaac_ros-dev/src/module89 --rosdistro foxy -y\"" >> ~/.bashrc
+echo "alias build=\"cd /workspaces/isaac_ros-dev && colcon build --packages-select module89 \"" >> ~/.bashrc
+echo "alias ins=\"cd /workspaces/isaac_ros-dev && . /workspaces/isaac_ros-dev/install/setup.bash\"" >> ~/.bashrc
+source ~/.bashrc
 
 sudo apt-get update
 rosdep update
@@ -82,6 +85,7 @@ pip install pyclipper simplejson tensorflow-object-detection-api
 sudo apt install -y libv4l-dev
 ## Install python-v4l2 library from source
 cd ~ || exit 1
+sudo apt-get install -y libv4l-dev
 git clone https://github.com/jnohlgard/python-v4l2capture
 cd ~/python-v4l2capture || exit 1
 python3 setup.py build
@@ -100,34 +104,48 @@ echo "Fix executable permission of *.py"
 sudo chmod 777 -R /workspaces/isaac_ros-dev/
 # sudo chmod 777 /workspaces/isaac_ros-dev/install/module89/lib/module89/*.py
 
-# OpenCV (Build from source)
-echo "Build OpenCV from source"
-export VERSION=4.2.0
-cd ~ || exit 1
-wget -O opencv.zip https://github.com/opencv/opencv/archive/$VERSION.zip
-wget -O opencv_contrib.zip https://github.com/opencv/opencv_contrib/archive/$VERSION.zip
-unzip opencv.zip
-unzip opencv_contrib.zip
-mv opencv-$VERSION opencv
-mv opencv_contrib-$VERSION opencv_contrib
-rm opencv.zip
-rm opencv_contrib.zip
-wget https://bootstrap.pypa.io/get-pip.py
-sudo python3 get-pip.py
-pip install numpy
-cd opencv
-mkdir build && cd build
-cmake -D CMAKE_BUILD_TYPE=RELEASE \
-    -D CMAKE_INSTALL_PREFIX=/usr/local \
-    -D WITH_CUDA=OFF \
-    -D INSTALL_PYTHON_EXAMPLES=ON \
-    -D INSTALL_C_EXAMPLES=ON \
-    -D OPENCV_EXTRA_MODULES_PATH=~/opencv_contrib/modules \
-    -D OPENCV_ENABLE_NONFREE=ON \
-    -D PYTHON_EXECUTABLE=/usr/bin/python3 \
-    -D BUILD_EXAMPLES=ON ..
-make -j$(nproc)
-sudo make install
-sudo ldconfig
+arch=$(uname -i)
+if [[ $arch == aarch64 ]]; then
+  # OpenCV (Build from source) [arrch64 only]
+  echo "Build OpenCV from source"
+  export VERSION=4.2.0
+  cd ~ || exit 1
+  wget -O opencv.zip https://github.com/opencv/opencv/archive/$VERSION.zip
+  wget -O opencv_contrib.zip https://github.com/opencv/opencv_contrib/archive/$VERSION.zip
+  unzip opencv.zip
+  unzip opencv_contrib.zip
+  mv opencv-$VERSION opencv
+  mv opencv_contrib-$VERSION opencv_contrib
+  rm opencv.zip
+  rm opencv_contrib.zip
+  wget https://bootstrap.pypa.io/get-pip.py
+  sudo python3 get-pip.py
+  pip install numpy imutils
+  cd opencv
+  mkdir build && cd build
+  cmake -D CMAKE_BUILD_TYPE=RELEASE \
+      -D CMAKE_INSTALL_PREFIX=/usr/local \
+      -D WITH_CUDA=OFF \
+      -D INSTALL_PYTHON_EXAMPLES=ON \
+      -D INSTALL_C_EXAMPLES=ON \
+      -D OPENCV_EXTRA_MODULES_PATH=~/opencv_contrib/modules \
+      -D OPENCV_ENABLE_NONFREE=ON \
+      -D PYTHON_EXECUTABLE=/usr/bin/python3 \
+      -D BUILD_EXAMPLES=ON ..
+  make -j$(nproc)
+  sudo make install
+  sudo ldconfig
+fi
+
+# Setup DOPE
+echo "Build & Install DOPE"
+cd /workspaces/isaac_ros-dev
+colcon build --packages-up-to isaac_ros_dope
+chmod +x /workspaces/isaac_ros-dev/install/setup.bash
+. /workspaces/isaac_ros-dev/install/setup.bash
+echo "Build & Install module89"
+rosdep install -i --from-path /workspaces/isaac_ros-dev/src/module89 --rosdistro foxy -y
+colcon build --packages-up-to module89
+. /workspaces/isaac_ros-dev/install/setup.bash
 
 $@
